Added AIV Semester 3 work
Semester 3 is pretty much semester 2 upgraded. A lot of the material is
the same, but done more accurately. I was still not coding at an elegant
level by any means, but it was this point that I started to comprehend
the course material, instead of simply perform the actions in order to
pass.