Added ".gitignore" file, removed unnecessary Semester_1 files, and added ignored files.
I added a ".gitignore" file so that I can include all the files I'd like
to keep, but do not want publicly shared on the internet, stored in the
same place, so that I do not have to micromanage a duplicate local
repository with no git files simply to keep a logical storage pattern of
data.
I went through the "Semester_1" folder and removed all unnecessary
files.
I added ignored files into the "Semester_1" folder.
I also changed the location of some files.