Added AIV Non-Semester Specific gitignore files, Up to date on semester specific gitignorefiles, Added VGP Program Guide
Added all folders from the "Art Institute of Vancouver" that didn't
belong to a specific semester into the folder "Non-Semester Specific".
All of these are gitignore files with the exception of the file "VGP
Program Guide.pdf" which resides in the "General" folder.
Added all files from school classes that should are gitignore
Added "Multiple Locations" and "File Types" to .gitignore, as well as
reformatted the entire file.