Added wk3 in class changes (performed on "mine"). Removed debug logs.
In class we modified some of the files by following along the teacher's
work on our own version of the engine. This is why the changes are
reflected in the "mine" folder and not the "teacher" folder.

I also removed unnecessary debug log files that were previously
committed by accident.