Completed assignment 1 homework up to DrawLine with colour, but not implemented into script parser.
Completed assignment 1 homework up to DrawLine with colour, but not
implemented into script parser, and thus needs more work before it can
be considered ready for hand in.
I am committing now so that I can continue working on this at school.
I also moved the majority of the CMGfx engine files into folders that
correspond to their named filters in Visual Studio.