Implemented everything required up to end of Assignment 2. Line drawing needs work.
Everything from the Assignment 1 and 2 documents is now coded, but
there is still some checking that needs to be done.
The number one issue is calling "drawbegin line" correctly draws the
end points of the line along with their corresponding colours, but does
not draw the pixels in between.
Note: Assignment 1 got turned into Assignments 1,2 and 3 all in one,
thus this branch which was supposed to be purely for assignment 1 is
still technically correctly titled, although confusing.