Triangles now rendering. Vertical line bug fixing still required.
Using a combination of the Duong Hoac's code and mine, I have gotten
triangles to correctly render, with the exception of random vertical
lines simply refusing to draw.
I am committing now as the entire project builds, and this vertical line
issue is only a non-critical run-time bug.
This commit is still on the Assignment1 branch as the time allotted to
complete the assignment has been extended for the entire class, but as a
result, assignment 1 now involves assignments 1-3 all in one for real
this time.