Triangle vertical line bug fixed. Clipping implemented.
Not tested, but seems to be working perfectly. Class starting in 5min,
so I'm checking this in now. I think I am completely done, but I plan to
return to this for error checking and upgrading going forward, so that I
can safely merge this back into master to use for future assignments.